# crestdb-data/src/gen/java/hep/crest/swagger/model/TagSetDto.java
/*
 * CrestDB REST API
 * OpenApi3 for conditions data management
 *
 * The version of the OpenAPI document: 2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package hep.crest.swagger.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import hep.crest.swagger.model.CrestBaseResponse;
import hep.crest.swagger.model.GenericMap;
import hep.crest.swagger.model.TagDto;
import hep.crest.swagger.model.TagSetDtoAllOf;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.util.ArrayList;
import java.util.List;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * An Set containing TagDto objects.
 */
@ApiModel(description = "An Set containing TagDto objects.")
@JsonPropertyOrder({
  TagSetDto.JSON_PROPERTY_FORMAT,
  TagSetDto.JSON_PROPERTY_RESOURCES
})
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaJerseyServerCodegen")
public class TagSetDto extends CrestBaseResponse  {
  public static final String JSON_PROPERTY_FORMAT = "format";
  @JsonProperty(JSON_PROPERTY_FORMAT)
  private String format = "TagSetDto";

  public static final String JSON_PROPERTY_RESOURCES = "resources";
  @JsonProperty(JSON_PROPERTY_RESOURCES)
  private List<TagDto> resources = null;

  public TagSetDto format(String format) {
    this.format = format;
    return this;
  }

  /**
   * Get format
   * @return format
   **/
  @JsonProperty("format")
  @ApiModelProperty(value = "")
  
  public String getFormat() {
    return format;
  }

  public void setFormat(String format) {
    this.format = format;
  }

  public TagSetDto resources(List<TagDto> resources) {
    this.resources = resources;
    return this;
  }

  public TagSetDto addResourcesItem(TagDto resourcesItem) {
    if (this.resources == null) {
      this.resources = new ArrayList<>();
    }
    this.resources.add(resourcesItem);
    return this;
  }

  /**
   * Get resources
   * @return resources
   **/
  @JsonProperty("resources")
  @ApiModelProperty(value = "")
  @Valid 
  public List<TagDto> getResources() {
    return resources;
  }

  public void setResources(List<TagDto> resources) {
    this.resources = resources;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    TagSetDto tagSetDto = (TagSetDto) o;
    return Objects.equals(this.format, tagSetDto.format) &&
        Objects.equals(this.resources, tagSetDto.resources) &&
        super.equals(o);
  }

  @Override
  public int hashCode() {
    return Objects.hash(format, resources, super.hashCode());
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class TagSetDto {\n");
    sb.append("    ").append(toIndentedString(super.toString())).append("\n");
    sb.append("    format: ").append(toIndentedString(format)).append("\n");
    sb.append("    resources: ").append(toIndentedString(resources)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}