# swagger_schemas/swagger/yaml/crestdb_full.yaml
openapi: 3.0.0
info:
  version: "2.0"
  title: "CrestDB REST API"
  description: "OpenApi3 for conditions data management"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
  /admin/globaltags/{name}:
    put:
      tags:
        - admin
      summary: Update a GlobalTag in the database.
      description: 'This method allows to update a GlobalTag.Arguments: the name has to uniquely identify a global tag.'
      operationId: updateGlobalTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalTagDto'
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalTagDto"
    delete:
      tags:
        - admin
      summary: Remove a GlobalTag from the database.
      description: 'This method allows to remove a GlobalTag.Arguments: the name has to uniquely identify a global tag.'
      operationId: removeGlobalTag
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
...
  /fs/tar:
    post:
      tags:
        - fs
      summary: Dump a tag into filesystem and retrieve the tar file asynchronously.
      description: This method allows to request a tar file from the server using a tag specified in input.
      operationId: buildTar
      parameters:
        - name: tagname
          in: query
          description: 'tagname: the tag name {none}'
          required: true
          schema:
            type: string
            default: none
        - name: snapshot
          in: query
          description: 'snapshot: the snapshot time {0}'
          required: true
          schema:
            type: integer
            default: 0
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
  /fs/tag:
    post:
      tags:
        - fs
      summary: Read a tag from filesystem. If the tag does not exists dump it.
      description: This method allows to dump a tag on filesystem.
      operationId: findFsTag
      parameters:
        - name: tagname
          in: query
          description: 'tagname: the tag name {none}'
          required: true
          schema:
            type: string
            default: none
        - name: reqid
          in: query
          description: 'reqid: the request id from tar method {none}'
          required: true
          schema:
            type: string
            default: none
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagSetDto"
...
  /globaltags:
    get:
      tags:
        - globaltags
      summary: Finds a GlobalTagDtos lists.
      description: 'This method allows to perform search and sorting.Arguments: by=<pattern>, page={ipage}, size={isize}, sort=<sortpattern>. The pattern <pattern> is in the form <param-name><operation><param-value>       <param-name> is the name of one of the fields in the dto       <operation> can be [< : >] ; for string use only [:]        <param-value> depends on the chosen parameter. A list of this criteria can be provided       using comma separated strings for <pattern>.      The pattern <sortpattern> is <field>:[DESC|ASC]'
      operationId: listGlobalTags
      parameters:
        - name: by
          in: query
          description: 'by: the search pattern {none}'
          required: false
          schema:
            type: string
            default: none
        - name: page
          in: query
          description: 'page: the page number {0}'
          required: false
          schema:
            type: integer
            default: 0
            format: int32
        - name: size
          in: query
          description: 'size: the page size {1000}'
          required: false
          schema:
            type: integer
            default: 1000
            format: int32
        - name: sort
          in: query
          description: 'sort: the sort pattern {name:ASC}'
          required: false
          schema:
            type: string
            default: 'name:ASC'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalTagSetDto"
    post:
      tags:
        - globaltags
      summary: Create a GlobalTag in the database.
      description: 'This method allows to insert a GlobalTag.Arguments: GlobalTagDto should be provided in the body as a JSON file.'
      operationId: createGlobalTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalTagDto'
      parameters:
        - name: force
          in: query
          description: 'force: tell the server if it should use or not the insertion time provided {default: false}'
          required: false
          schema:
            type: string
            default: false
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalTagDto"
  /globaltags/{name}:
    get:
      tags:
        - globaltags
      summary: Finds a GlobalTagDto by name
      description: This method will search for a global tag with the given name. Only one global tag should be returned.
      operationId: findGlobalTag
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GlobalTagSetDto"
  /globaltags/{name}/tags:
    get:
      tags:
        - globaltags
      summary: Finds a TagDtos lists associated to the global tag name in input.
      description: 'This method allows to trace a global tag.Arguments: record=<record> filter output by record, label=<label> filter output by label'
      operationId: findGlobalTagFetchTags
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: record
          in: query
          description: 'record:  the record string {}'
          required: false
          schema:
            type: string
            default: none
        - name: label
          in: query
          description: 'label:  the label string {}'
          required: false
          schema:
            type: string
            default: none
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagSetDto"
...
  /tags/{name}:
    get:
      tags:
        - tags
      summary: Finds a TagDto by name
      description: This method will search for a tag with the given name. Only one tag should be returned.
      operationId: findTag
      parameters:
        - name: name
          in: path
          description: 'name: the tag name'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagSetDto"
            application/xml:
              schema:
                $ref: "#/components/schemas/TagSetDto"
    put:
      tags:
        - tags
      summary: Update a TagDto by name
      description: 'This method will search for a tag with the given name, and update its content for the provided body fields. Only the following fields can be updated: description, timeType, objectTime, endOfValidity, lastValidatedTime.'
      operationId: updateTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericMap'
      parameters:
        - name: name
          in: path
          description: 'name: the tag name'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
            application/xml:
              schema:
                $ref: "#/components/schemas/TagDto"
...
servers:
  - url: http://localhost:8080/api
components:
  schemas:
...
    CrestBaseResponse:
      type: object
      properties:
        size:
          type: integer
          format: int64
        datatype:
          type: string
        filter:
          $ref: '#/components/schemas/GenericMap'
      discriminator:
        propertyName: format
    GenericMap:
      type: object
      properties:
        name:
          type: string
      additionalProperties:
        type: string
...
    GlobalTagDto:
      type: object
      properties:
        name:
          type: string
        validity:
          type: number
        description:
          type: string
        release:
          type: string
        insertionTime:
          type: string
          format: date-time
        snapshotTime:
          type: string
          format: date-time
        scenario:
          type: string
        workflow:
          type: string
        type:
          type: string
        snapshotTimeMilli:
          type: integer
          format: int64
        insertionTimeMilli:
          type: integer
          format: int64
...
    TagDto:
      type: object
      properties:
        name:
          type: string
        timeType:
          type: string
        payloadSpec:
          type: string
        synchronization:
          type: string
        description:
          type: string
        lastValidatedTime:
          type: number
        endOfValidity:
          type: number
        insertionTime:
          type: string
          format: date-time
        modificationTime:
          type: string
          format: date-time
...
    TagSetDto:
      description: An Set containing TagDto objects.
      allOf:
        - $ref: '#/components/schemas/CrestBaseResponse'
        - properties:
            format:
              type: string
              default: TagSetDto
            resources:
              type: array
              items:
                $ref: '#/components/schemas/TagDto'
...
    GlobalTagSetDto:
      description: An GlobalTagSet containing GlobalTagDto objects.
      allOf:
        - $ref: '#/components/schemas/CrestBaseResponse'
        - properties:
            format:
              type: string
              default: GlobalTagSetDto
            resources:
              type: array
              items:
                $ref: '#/components/schemas/GlobalTagDto'
...